backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
